requirements.txt
Flask
Flask-SocketIO
fitz
Pillow
pytesseract
opencv-python
pandas
Werkzeug
PyMuPDF
openpyxl